\name{support.BWS3-package}
\alias{support.BWS3-package}
\alias{support.BWS3}
\docType{package}

\title{
Tools for Case 3 best-worst scaling
}

\description{
Provides basic functions that support an implementation of Case 3 (multi-profile case) best-worst scaling.
}

\details{
Case 3 (multi-profile case) best-worst scaling (BWS) is a question-based survey method to elicit people's preferences for attribute levels. A Case 3 BWS question consists of three or more profiles. Each profile has two or more attributes and each attribute has two or more levels. Respondents are asked to select the best and worst profiles for each question. Analyzing the responses to the questions reveals the people's preferences for the attribute levels. Refer to Hensher et al. (2015), Lancsar et al. (2013), Louviere et al. (2015), and Marley and Pihlens (2012) for details. 

\bold{Acknowledgments}

I would like to thank Professor Kazuo Sato for his kind support. This work was supported by JSPS KAKENHI Grant Number JP16K07886.
}

\author{
Hideo Aizaki

Maintainer: Hideo Aizaki <azk-r@spa.nifty.com>
}

\references{
Hensher DA, Rose JM, Greene WH (2015) \emph{Applied Choice Analysis}. 2nd edition. Cambridge University Press. \url{https://doi.org/10.1017/CBO9781316136232}.

Louviere JJ, Flynn TN, Marley AAJ (2015) \emph{Best-Worst Scaling: Theory, Methods and Applications}. Cambridge University Press. \url{https://doi.org/10.1017/CBO9781107337855}.

Lancsar E, Louviere J, Donaldson C, Currie G, Burgess L (2013) Best worst discrete choice experiments in health: Methods and an application. \emph{Social Science & Medicine} \bold{76}: 74--82. \url{https://doi.org/10.1016/j.socscimed.2012.10.007}.

Marley AAJ, Pihlens D (2012) Models of best-worst choice and ranking among multiattribute options (profiles). \emph{Journal of Mathematical Psychology} \bold{56}: 24--34. \url{https://doi.org/10.1016/j.jmp.2011.09.001}.

}

\keyword{ package }

